feat: product card에서 card info 분리
