+++ src/components/molecules/ProductCard.tsx
@@ -1,7 +1,6 @@
+import CardInfo from '@components/atoms/card/CardInfo';
 import ImageWithCheck from '@components/atoms/card/ImageWithCheck';
 import ImageWithHeart from '@components/atoms/card/ImageWithHeart';
-import PaybackPrice from '@components/atoms/card/PaybackPrice';
-import { getPriceText } from '@utils/formatData';
 
 type ProductCardProps = {
 	thumbnail: string;
@@ -11,18 +10,16 @@ type ProductCardProps = {
 	size: 'small' | 'medium' | 'large';
 	type: 'default' | 'checkbox' | 'heart';
 	isFullHeart?: boolean;
-	category?: string;
 };
 
 const ProductCard = ({
 	thumbnail,
 	title,
 	price,
-	paybackPrice,
 	type,
+	paybackPrice,
 	isFullHeart,
 	size,
-	category,
 }: ProductCardProps) => {
 	const Image = {
 		checkbox: <ImageWithCheck src={thumbnail} alt={title} />,
@@ -47,52 +44,25 @@ const ProductCard = ({
 
 	const stylesBySize = {
 		small: {
-			title: 'typography-Body4 typography-M line-clamp-1',
-			price: 'typography-Body2 typography-M',
-			pricesContainer: 'flex flex-col gap-1',
 			containerGap: 'gap-1',
 		},
 		medium: {
-			title: 'typography-Body3 typography-M line-clamp-1',
-			price: 'typography-Body1 typography-M',
-			pricesContainer: 'flex flex-col gap-1',
 			containerGap: 'gap-1',
 		},
 		large: {
-			title: 'typography-Body1 typography-R',
-			price: 'typography-Subhead',
-			pricesContainer: 'flex gap-6',
 			containerGap: 'gap-5',
 		},
 	};
 
-	const renderCategory = () => {
-		if (!category) {
-			null;
-		}
-
-		return (
-			<span className="typography-Body4 typography-R text-Secondary_B">
-				{category}
-			</span>
-		);
-	};
-
 	return (
-		<div
-			className={`flex flex-col ${stylesBySize[size].containerGap} text-White`}
-		>
+		<div className={`flex flex-col ${stylesBySize[size].containerGap}`}>
 			{Image[type]}
-			<div>
-				{renderCategory()}
-				<span className={stylesBySize[size].title}>{title}</span>
-				<div className={stylesBySize[size].pricesContainer}>
-					<span className={stylesBySize[size].price}>
-						{getPriceText(price)}
-					</span>
-					<PaybackPrice paybackPrice={paybackPrice} size={size} />
-				</div>
-			</div>
+			<CardInfo
+				title={title}
+				price={price}
+				paybackPrice={paybackPrice}
+				size={size}
+			/>
 		</div>
 	);
 };
